+++ bin/c-d.js
@@ -6,6 +6,7 @@ const program = require('commander');
 
 // The basic program, which uses sub-commands.
 program
-    .command('dc <command>', 'Shortcuts to control Docker Compose.')
-    .command('d <command>', 'Shortcuts to control Docker.')
+    .command('clean <type>', 'Removes unused containers and images')
+    .command('images', 'List images')
+    .command('ps', 'List containers')
     .parse(process.argv);